project/requirements.txt
joblib~=1.1.0
librosa~=0.8.1
numpy~=1.19.2
pandas~=1.3.4
PyYAML~=6.0
scikit_learn~=1.0.1
tensorflow~=2.4.0
tqdm~=4.57.0
pyyaml